More uniform font rendering in overlay
Fixes an issue where overlay text would be always scaled to fit a
given height, leading to uneven sizes depending on content

As long as text is within font metrics, it will not be scaled. Some
extremely high characters (accents on capitals, Zalgo, etc.) will
however be scaled to fit a bounding height, while preserving the
baseline to avoid breaking the regular layout.