Fix overlay causing stack overflow when closing vs2010 solution
The amount of stack allocated for reading the black/whitelist values from the registry caused a stack overflow when closing solutions in vs2010. The fix is to switch to dynamic allocation on the heap for this variable. This fix also contains minor refactoring for better readability and adds a bit of additional debug output.