Add a workaround for broken CodecVersion messages on 1.2.2 servers
With this patch the client's codec versions should match the versions
on the server.

The 1.2.2 had a bug which caused it to send the same version for alpha
and beta, but we know that the codec switching code will always change
only one codec version, not both and that it doesn't know the correct
bitstream constant for CELT 0.7.x, therefore it will always invert the
state of preferAlpha.

We can use that to decide which codec to change on the client and reset
the other version to the version the client already had in that slot.