Work around scrolling behavior of QTextBrowser in Log(Fixes #3298338).
QTextBrowser uses a top anchor for its scrolling behavior. As such resizing the Widget from the bottom makes it "scroll" away from the bottom of the log if the widget get smaller. This behavior interacted badly with the new scrolling behavior as auto-scrolling would get disabled if such a resize occurred. As a workaround the log now automatically scrolls to bottom if resized.