Fix: Stale custom context menu entries (#3301658)
Added a new ContextActionRemove protobuf message type triggering the clientside removal of a context menu entry. Since the server only remembers the action identifier and doesn't have a notion of where it is applied to on the client it is only possible to remove all entries for an identifier at the same time at this point. The removal is triggered when overriding a new handler (clientside visibility might have changed), remove a handler and in error conditions.