Fix iObfuscate for IPv6
IPv6-mapped IPv4 addresses contain to many zeros to
use XOR without revealing the "key". Switching the
output to a sha1 hash of the random value + the address
solves the problem. We don't include the port in the hash
because it's useful to see which connection attempts were
from the same IP address even if you can't read the actual
value.