Part 1 of 2; Implementing volume restoration.
This part of the fix reworks how sinks are stored internally and how we
restore the volumes of them. We keep track of some new properties, so
that we don't have to rely on just a stream ID. This has the nice
benefit that if an application is closed an restarted during
attenuation, we'll still recognize it and fix the volume.

Additionally, we no longer mangle the volume if it was changed during
attenuation by the user.

What is *not* implemented yet is if a sink input (application) dies or
exits, the volume will be restored. This commit only deals with streams
that are still alive at the end of attenuation.