Part 2 of 2: Fix volumes in stream restore db.
This handles the issue mentioned in the previous commit: when a
sink/application terminates, we no longer can change its volume
directly. So, using PulseAudio's module-stream-restore, we save the
fixed volume in the database manually.

At this point, Pulse attenuation should be pretty robust, working
properly in cases with normal, disappearing, changed, and reappearing
sink inputs.