Restore volumes unconditionally on exit.
This was a bit of a pain. Because all of the attenuation code is running
in the PulseAudio event thread, we *cannot* make the volume restoration
synchronous. Additionally, Pulse doesn't provide a way to wait for all
events to complete, which is needed so that volume changes take effect.

So, we keep track of the number of pending operations (volume
adjustments and restore database writes) during shutdown. Once it hits
zero, ping the main thread, and finish shutting down. Consider this a
hack, but it cannot be fixed without a significant rewrite of the
threading code, which I'd like to avoid. :)