Fix ghost disconnect removing permissions (#3423985)
Returning a list of groups using an authenticator added
temporary group membership for the user id only. Those are
removed when we're disconnecting an old user using the same
user id. Adding the session id to the group should fix the
problem.

This won't work if someone uses verifyPassword to set the
permissions (which is an undocumented implementation detail
anyway).