Fix strong cert check for self-signed server certificates
If the server is using a self-signed certificate the strong
certificate check for the client cert will fail with a
"The certificate is self-signed" error even if it's a valid
cert. It looks like Qt is propagating the "error" for the server
certificate to the client cert check. Adding the server certificate
as a CA certificate fixes the problem.