fix bug 924: Server favorites disappear on second Mumble instance (-m)
On Windows, using journal mode truncate does not seem to work, at least with our Qt / Qt 4.8.
The issue is not limited to the explicit transaction, but occurs even after just the DELETE FROM `servers` query, without a transaction, and also as not a prepared statement.
It does not occur with a single instance accessing the DB, but when more than one are connected to the DB / the DB is opened anew.
Using journal_mode DELETE instead of our optimization to use TRUNCATE fixes the issue. As it seems to be limited to windows, for windows we now use DELETE, and for the rest we still use TRUNCATE.