Updated L4D2 Offsets
Updated the L4D2 Offsets.  Most of them were just small shifts in
memory.

The state offset no longer appears to work however so I simply
changed it so state is derived from the hostname (which is an empty
string in menus).  This may lead to a brief period during loading where
the hostname is set but the positions aren't yet, but it should only be
momentary.