Fix application shortcuts not working in minimal mode.
Hiding elements of the UI like the icon toolbar disabled the
actions in it. As we want the shortcuts of the actions always
available we need to work around this behavior.

This patch adds all actions found in the object tree of the main-window
directly to the main window. As we don't use the auto-generated
context menu for that there shouldn't be any side-effects.

Another possible way would be to manually create QShortcuts and use
those instead. This would have much bigger code impact though.