compiler.pri: use default OS X SDK instead of hard-coding it.
For Qt 5, we simply use 'macosx'. For Qt 4, we query via 'xcrun',
just like Qt 5 does under the covers.

This allows us to avoid changing the hard-coded OS X SDK every time
a new SDK is released. As long as Xcode is up-to-date, we'll build
with the new SDK automatically.