mumble, mumble_exe: fix overlay self-detection for the mumble_app.dll build.
When Mumble was built as the mumble.exe/mumble_app.dll pair, mumble_ol.dll's
overlay self-detection logic didn't work, because it failed to find the
special "mumbleSelfDetection" symbol.

This change adds that symbol to mumble_exe. It is perhaps not the cleanest
solution to the problem, but doing it this way allows older mumble_ol.dlls
to continue to work with mumble_app.dll builds, and we avoid changing the
overlay self-detection check in mumble_ol.dll.

Fixes issue #1053.