Fix accidental CBR limit for large Opus frames.
With Opus frames are no longer limited to a duration of 10ms
like with CELT. This means a single frame can now become
bigger than previously possible. At our maximum supported rate
of 96kbit/s with 60ms frames a single frame will be 720 bytes.
Previously the encoding output buffer was 512 bytes which triggered
rate limiting in Opus to hit this goal.

This patch bumps the buffer size to 960 and uses boost::array instead
of a C-Style array to ensure future changes don't accidentally miss
places where the size of the buffer is relevant.