TextToSpeech_macx: selectively conform to NSSpeechSynthesizerDelegate protocol.
MUSpeechSynthesizerPrivateHelper currently conforms to the
NSSpeechSynthesizerDelegate protocol, which was an informal
protocol prior to 10.6.

Our legacy universal builds currently use the ancient 10.5 SDK
and does not know of the NSSpeechSynthesizerDelegate protocol,
which was first introduced in the 10.6 SDK. Because of that, the
legacy universal build is currently broken.

To fix it, we change the code to only adhere to
NSSpeechSynthesizerDelegate via a class extension. This class
extension is only exposed to non-universal builds using the 10.6
SDK (or greater).

(cherry picked from commit 1bd57bd0a56c6eed4c7cf3fd477ca575c178f307)