Add hidden option to disable pubserver list (FR #1120)
Added ui/disablepubliclist which if set to true removes the
public server list in the connect dialog. This option isn't
available in the UI and is meant for customizing setups in
specific deployments.

On windows you can enable the option by creating a string key
disablepubliclist under HKEY_CURRENT_USER\Software\Mumble\Mumble\ui
with the value true.