Drastically simplify release.pl.
Previously release.pl did complex parsing of .pro
files to figure out which files to include in the
tar ball.

This change removes this logic and instead performs
the minimal modifications needed under the assumption
of a clean tree and then uses system tools to create
the tar.gz and zip files.