Only attenuate on same sink
Until this patch, Mumble would attenuate all input sources to all sinks. Rarely, this is undesirable behavior.

Scenario: Live streaming a particular application, but running Mumble off-stream. Two null-sinks are created, one called "stream_spkr", with a loopback to the speakers. The other, "stream", with a loopback coming from "stream_spkr". Without this patch, if Mumble is run on "stream_spkr" or the hardware output sink, and another audio program is run on "stream" (which has no inputs into the "stream_spkr" or hardware output sinks), Mumble will attenuate the music application being run on the "stream" sink.

I believe that is is undesired behavior, so this patch fixes this. With this patch, Mumble only attenuates applications that are running on the same output sink as Mumble (or one further up the chain that has an input into Mumble's output sink). This effectively allows for the user to "whitelist" applications to keep them from being attenuated by running them on a separate sink.

This should not effect 99.9% of Mumble users, as the default PulseAudio setup is usually only one hardware sink and everything uses it. However, it is a boon to those of us with a custom PulseAudio setup and who would like to keep certain applications from being attenuated by Mumble.