Plugins: add support for the 'path' attribute in plugin auto-update XML.
This new attribute specifies an absolute path on a Mumble services
host that a given plugin is available at:

  <plugin name="borderlands2.dll"
          hash="(...)"
          path="/plugins/<os>/<abi>/borderlands2.dll" />

This path is directly usable with our WebFetch class.

We need this to be able to properly service multiple plugin ABIs
and/or OSes at the same time.

We could have overloaded the meaning of the 'name' attribute, and
have passed the absolute path in there, but the chosen approach
is more explicit about its intentions.

Being explicit about the 'path' attribute also makes it easier to
understand how backwards compatibility is handled: Since Mumble 1.2.5
and prior verisons do not know about the 'path' attribute, they will
keep simply using the 'name' attribute as their pointer to the DLL
file. This means we'll have to keep serving the DLLs for these older
clients at <webroot>/plugins/<name>.

In practice, this means that all plugins that can be requested by
these older clients will need to have a copy that lives at
<webroot>/plugins/<name>. Since these clients only support
(abi=1600, os=Win32) we can ensure on the backend that plugins for
that combo are served at /plugins/<name>.