Overlay: Add Mutex to doPresent where possible
* doPresent searches for and uses DevState from devMap.
Thus, it should actually Mutex the entire method.
As this is not possible because of another (potential) Mutex down the code
path, add the Mutex where possible.

Doing it this way has a performance cost with unknown advantage. Also, the
DevState potentially disappearing while the method is being executed is
pretty stupid. But it is the best we can do right now.