Remove flawed MX host existence check for cert wizard (#1178)
The certificate wizard tried to verify the host of email addresses
input by users by resolving their A record. This is flawed as
a host isn't required to have an A record for receiving mail. The
correct way would be to check the MX record instead. This isn't
possible in Qt 4. As the feature is of dubious value anyways this
patch simply removes it. Now every email matching (.+)@(.+) is
accepted.