Add translation update script with transifex issue workaround (#1195)
Script will do an lupdate run on mumble_en.ts, apply the duplication
of numerus tags workaround for the transifex missing string issue
(see #1195) and then create a commit with the updated translation
file. This file - at least for now - is meant to be run manually.
Make sure to check the created commit before pushing.