mumble: fix Mumble-SA-2014-006.
Usernames and channel names were not properly HTML-escaped
when used in Qt widgets that are rich-text enabled.

This commit fixes that, but also touches various other
similar cases where an escaped version is appropriate.

This commit is based on the following commits from the
master branch (Mumble 1.3.0):

  b7d9387bd6dacbad0b2345f03dd8502a51c42f6a
  1caaec763b91c8a12b11d7ceb37e21622f4da76e
  a0ebded7df388ce625dfbeb0c2bd65fb782da2a4
  c52dedce8409da0654fd197690ff95411df3f9b2
  1f6ddaf35f9583ef72f2d924b76b425ec85fc692
  73a1a98d16a29c78d85e1c19b8feac7ba22dddfe
  d58990c374502cfdd4a5f2bd8dbd490f19e9e511
  b6e17cac696396d5bcbc3391846f9cacb9c072c4
  9837c4dc2d1d6c60505f7246cd00ffa33ad808a9
  17fa695b222a8b308438f9283e78dfba931e8712
  d9ff1e947d4d00f6cd8d38fd04cd87f6e3167028

Special thanks to Tim Cooper for various of the above
patches.