Update World of Warcraft Plugin to build 18291
I'm also tentatively throwing away the process of going through NameStore
looking for our characer name - all the posters on OwnedCore are doing it through
a simple pointer. I'm guessing this plugin doesn't do it that way because there was
no static pointer when it was written, but for now this seems to work.

(Also, no one has posted the updated offset for NameStorePtr yet)