PulseAudio: fix access to protected enums in AudioInput and AudioOutput.
The SampleFormat enums in AudioOutput and AudioInput are meant to be
used in subclasses of AudioOuptut and AudioInput, and because of that,
they're marked protected.

PulseAudioSystem has used these protected enums for some time, without
triggering any warnings. However, recent versions of Clang error out
on this behavior (and rightly so).

To fix this issue, this change modifies PulseAudioSystem to access the
SampleFormat enums via PulseAudioOutput and PulseAudioInput, instead
of going through AudioOutput and AudioInput directly. PulseAudioSystem
is already a friend class of PulseAudioOutput and PulseAudioInput, so
it can access the protected members of AudioOutput of AudioInput just
fine, as long as it happens via PulseAudioOutput and PulseAudioInput.

Originally reported on the Debian BTS (as #753273) by
Alexander <sanek23994@gmail.com>

Reported on the Mumble issue tracker (as #1302) by
Chris Knadle <Chris.Knadle@coredump.us>

(cherry picked from commit 01a5e83b76d049ef6ac44c8ce8c4f2cbd5f80e9e)