SRV records as in mumble-voip/mumble#1242
Please note that I'm not familiar with either the mumble codebase or Qt.
And because I was too lazy to set up a full Qt Windows devenv, I didn't
even try to compile this.

And I have no idea whether the retry/cycle-through-srv-records part could
work. And the server list probably needs similar code.

So now that we know that this commit is probably entirely broken and shit,
why did I do it? Because it clearly shows that adding support for SRV
records to mumble, a feature I would really appreciate, is trivial.