OSInfo: use 'WinX64' as the OS identifier for Windows/amd64 builds.
The OS string from OSInfo is used in places such as log files, and
in the Server Info and User Info dialogs.

I think it'll be beneficial in the long run to be able to
distinguish an x86 build from an amd64 on in these places.

Technically, the full OSInfo dump has an is64bit flag, but it's not
visible in the aforementioned places - only the OS string is. The
isb64bit flag is only used when submitting usage stats.