Make the Minimal View a Qt::Window, not a Qt::Tool
Making the Minimal View a tool window (Qt::Tool) stops the window from
showing up in the taskbar and makes it impossible to use ALT-TAB to get
to the window. Instead, make the Minimal View a regular window, so that
it shows up in the taskbar, can be recalled using ALT-TAB, and has
minimise/maximise window decorations.