Fix FTBFS regression from dd7cc7caa7257d2d1525daebbcc070cbb0983915
AA_UseHighDpiPixmaps is an option developed in Qt5 and only recently backported to some Qt4 releases.

When building against Qt 4.8.6, dd7cc7caa7257d2d1525daebbcc070cbb0983915 introduces a FTBFS regression. This conditional fixes the regression in all build environments that were tested but could have more tightened down values for QT_VERSION and OS_VERSION with more testing. There are bug reports that this flag will also not work on MACOSX 10.6 and lower, but this was not tested.