Add poor man's padding to last column in connect dialog.
Adds a space after the content of the last column to
prevent it being to close to the border of the treewidget.
Unfortunately we cannot use QSS which would make this
trivial as that would prevent users from customizing this
value with their own styles. Also programmatic access
to the size of an item to correct the hint is quite
involved. The single space seems to be the best solution.

Fixes #1331