Fix recording timer not properly displaying in Qt 5 version of Mumble.
Apparently Qt 5 changed QTime's behavior with regards to invalidity.
A default constructed QTime is midnight time and invalid. We added
some milliseconds and converted to string. Both Qt 4 and 5 docs state
that an invalid time will yield an empty string. In Qt 4 however our
use did give us the time string we wanted instead of an empty string.
This patch simply removes the use of the default constructor in favor
of manually creating a valid midnight QTime to work with.

Fixes #1426